[FIX] Level fix fixed level being not setted by func

+++ Source/GamePrototype/Public/Characters/Components/XPComponent.h
@@ -19,7 +19,7 @@ public:
 
 	// ILevelInteraction
 	virtual int GetLevel() override;
-	virtual void SetLevel(float) override;
+	virtual void SetLevel(int) override;
 	virtual OnLevelUpSignature& OnLevelUp() override;
 
 protected:
